refactor: moved components into shared and layout folders
