+++ app/frontend/src/components/layout/Sidebar.tsx
