+++ app/frontend/src/components/shared/ReportCard.tsx
