+++ app/frontend/src/components/shared/nearbyReports.tsx
