+++ app/frontend/src/pages/users/under-construction.tsx
@@ -1,5 +1,5 @@
 import { UserSideBar } from "../../components/users/usersSidebar";
-import { UnderConstruction } from "../../components/underConstruction";
+import { UnderConstruction } from "../../components/shared/underConstruction";
 
 export default function RegisteredUnderConstruction(){
     return(
